+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 347
+#define BUILD 348